Update xPDO
Squashed commit of the following:

commit 5c5ee86f69a356e5fdc7fdc6761e7fa46066215c
Merge: de49cfb67 0c9f2f009
Author: Jason Coward <jason@opengeek.com>
Date:   Wed Jan 22 09:16:07 2020 -0700

    Merge remote-tracking branch 'xpdo/2.x' into 2.x

    * xpdo/2.x:
      Fix capitalization of XPDO reference to xPDO

commit 0c9f2f0090ce52f6f67bd545982e6e83fd825482
Merge: 002f0ed81 5f5bd3e72
Author: Jason Coward <jason@opengeek.com>
Date:   Wed Jan 22 09:15:20 2020 -0700

    Merge remote-tracking branch 'xpdo/2.x' into 2.x

    * xpdo/2.x:
      Fix capitalization of XPDO reference to xPDO

commit 5f5bd3e72c753d0f7d7a24160f2b3ebdb258014a
Author: Jason Coward <jason@opengeek.com>
Date:   Wed Jan 22 09:14:14 2020 -0700

    Fix capitalization of XPDO reference to xPDO

commit de49cfb671907e1a3939588855f464a9911e23ff
Merge: c2c615788 002f0ed81
Author: Jason Coward <jason@opengeek.com>
Date:   Wed Jan 22 09:12:14 2020 -0700

    Merge remote-tracking branch 'xpdo/2.x' into 2.x

    * xpdo/2.x:
      Fix the wrong date format being used when getting a datetime value #158
      Fix invalid reference to $rObj in merged PR #149
      Fix invalid datetime value being used when formatting with regex
      Fix xPDOFileCache->delete() when a directory with same cache key exists
      Restrict extracted files to allowed extensions by option
      Load related object attributes correctly into nested relations

commit 002f0ed81254689c5a38272dc7f47f2d0dedb955
Merge: f414d7c96 d4508ac5b
Author: Jason Coward <jason@opengeek.com>
Date:   Wed Jan 22 09:05:17 2020 -0700

    Merge remote-tracking branch 'xpdo/2.x' into 2.x

    * xpdo/2.x:
      Fix the wrong date format being used when getting a datetime value #158
      Fix invalid reference to $rObj in merged PR #149
      Fix invalid datetime value being used when formatting with regex
      Fix xPDOFileCache->delete() when a directory with same cache key exists
      Restrict extracted files to allowed extensions by option
      Load related object attributes correctly into nested relations

commit d4508ac5b5c2912308f0b42952b22fb780a7dc5d
Author: Jason Coward <jason@opengeek.com>
Date:   Thu Jan 16 09:03:46 2020 -0700

    Fix the wrong date format being used when getting a datetime value #158

    The first patch did not fully solve that issue

commit bd53a68dfe76e515d8fd04c8e388e18fbae37392
Author: Jason Coward <jason@opengeek.com>
Date:   Tue Jan 14 10:48:41 2020 -0700

    Fix invalid reference to $rObj in merged PR #149

commit 2e7d9361cfc19a97b679c5ea3c6124ee14f8f5aa
Merge: 03bc88ded c552690a6
Author: Jason Coward <jason@opengeek.com>
Date:   Tue Jan 14 10:13:57 2020 -0700

    Load related object attributes correctly into nested relations

    Merge remote-tracking branch 'origin/pr/149' into 2.x

    * origin/pr/149:
      Load related object attributes correctly into nested relations

commit 03bc88ded79646300c91bfd9229f02e26a239b19
Merge: c104be73d 590c05d76
Author: Jason Coward <jason@opengeek.com>
Date:   Sat Jan 11 18:47:19 2020 -0700

    Add option to restrict extraction of files by extension to xPDOZip

    Merge remote-tracking branch 'origin/pr/157' into 2.x

    * origin/pr/157:
      Restrict extracted files to allowed extensions by option

commit c104be73ddd0e4f254ff10f005c0b5c7c8b9efe4
Author: Jason Coward <jason@opengeek.com>
Date:   Sat Jan 11 18:31:50 2020 -0700

    Fix invalid datetime value being used when formatting with regex

    Resolves #158 in 2.x branch

commit 4d782c3d98c149ac74d03cef5c424632f261dbe4
Merge: 567cadb97 b09881338
Author: Jason Coward <jason@opengeek.com>
Date:   Thu Jan 9 12:22:05 2020 -0700

    Fix xPDOFileCache->delete() when a directory with same cache key exists

    Merge remote-tracking branch 'origin/pr/171' into 2.x

    * origin/pr/171:
      Fix xPDOFileCache->delete() when a directory with same cache key exists

commit b09881338a58361ea931e1235cdd0abe4ba9bafe
Author: Jason Coward <jason@opengeek.com>
Date:   Tue Dec 31 10:56:13 2019 -0700

    Fix xPDOFileCache->delete() when a directory with same cache key exists

    Other xPDOCache implementations check for multiple_object_delete option before attempting to recursively remove cache entries that match the cache key provided to the delete() method. xPDOFileCache was not doing this, and further, if a directory existed with the cache key to be deleted, the directory was emptied of cache entries but the actual cache entry that was meant to be deleted was left in place.

    For 2.x, cache entries in a folder matching the cache key will be removed recursively unless multiple_object_delete is explicitly set to false. This is to prevent a major change in behavior. For 3.x, it will be made consistent with other xPDOCache implementations and require multiple_object_delete to be explicitly set to true before entries in a folder matching the cache key will be removed.

    Resolves #170

commit 590c05d7644108c6e98a4904b1903ecad4aaefb5
Author: Thomas Jakobi <thomas.jakobi@partout.info>
Date:   Fri Mar 8 08:55:25 2019 +0100

    Restrict extracted files to allowed extensions by option

commit c552690a6540ba9700e8fa39ff7273c4dac4b1de
Author: Jan Peca <pecajan@gmail.com>
Date:   Fri Aug 31 16:02:41 2018 +0200

    Load related object attributes correctly into nested relations

+++ core/xpdo/cache/xpdocachemanager.class.php
@@ -1012,17 +1012,18 @@ class xPDOFileCache extends xPDOCache {
 
     public function delete($key, $options= array()) {
         $deleted= false;
-        $cacheKey= $this->getCacheKey($key, array_merge($options, array('cache_ext' => '')));
-        if (file_exists($cacheKey) && is_dir($cacheKey)) {
-            $results = $this->xpdo->cacheManager->deleteTree($cacheKey, array_merge(array('deleteTop' => false, 'skipDirs' => false, 'extensions' => array('.cache.php')), $options));
-            if ($results !== false) {
-                $deleted = true;
+        if ($this->getOption('multiple_object_delete', $options, true)) {
+            $cacheKey= $this->getCacheKey($key, array_merge($options, array('cache_ext' => '')));
+            if (file_exists($cacheKey) && is_dir($cacheKey)) {
+                $results = $this->xpdo->cacheManager->deleteTree($cacheKey, array_merge(array('deleteTop' => false, 'skipDirs' => false, 'extensions' => array('.cache.php')), $options));
+                if ($results !== false) {
+                    $deleted = true;
+                }
             }
-        } else {
-            $cacheKey= $this->getCacheKey($key, $options);
-            if (file_exists($cacheKey)) {
-                $deleted= @ unlink($cacheKey);
-            }
+        }
+        $cacheKey= $this->getCacheKey($key, $options);
+        if (file_exists($cacheKey)) {
+            $deleted= @ unlink($cacheKey);
         }
         return $deleted;
     }
